Light cache number of items as int

+++ lib/ethash/params.hpp
@@ -10,7 +10,7 @@ namespace ethash
 {
 constexpr static int light_cache_init_size = 1 << 24;
 constexpr static int light_cache_growth = 1 << 17;
-constexpr static size_t light_cache_rounds = 3;
+constexpr static int light_cache_rounds = 3;
 constexpr static int full_dataset_init_size = 1 << 30;
 constexpr static int full_dataset_growth = 1 << 23;
 constexpr static uint32_t full_dataset_item_parents = 256;
